fix fiber examples
