+++ docs/examples/get_started/1.state.html
@@ -8,22 +8,14 @@
 </head>
 <body>
   <div id="app"></div>
-  <script src="../../dist/index.js"></script>
+  <script src="kurge_1.1.0.js"></script>
   <script>
-    const { render, useState, createElement: h, useEffect } = Kurge
-
-    function useCount() {
-      const count = useState(0)
-      useEffect(() => {
-        document.title = 'Counter mounted'
-      }, [])
-      return count
-    }
+    const { render, useState, createElement: h } = Kurge
 
     function app() {
-      const count = useCount()
+      const count = useState(0)
       return h('div', {},
-        h('p', null, `You clicked ${count.value} times`),
+        h('p', null, 'You clicked ' + count.value + ' times'),
         h('button', { onClick: () => count.value++ }, 'Click me')
       )
     }
