+++ docs/examples/get_started/2.effect.html
@@ -8,7 +8,7 @@
 </head>
 <body>
   <div id="app"></div>
-  <script src="../../dist/index.js"></script>
+  <script src="kurge_1.1.0.js"></script>
   <script>
     const { render, useState, createElement: h, useEffect } = Kurge
