+++ docs/examples/get_started/5.custom-hook.html
@@ -8,14 +8,22 @@
 </head>
 <body>
   <div id="app"></div>
-  <script src="../../dist/index.js"></script>
+  <script src="kurge_1.1.0.js"></script>
   <script>
-    const { render, useState, createElement: h } = Kurge
+    const { render, useState, createElement: h, useEffect } = Kurge
+
+    function useCount() {
+      const count = useState(0)
+      useEffect(() => {
+        document.title = 'Counter mounted'
+      }, [])
+      return count
+    }
 
     function app() {
-      const count = useState(0)
+      const count = useCount()
       return h('div', {},
-        h('p', null, 'You clicked ' + count.value + ' times'),
+        h('p', null, `You clicked ${count.value} times`),
         h('button', { onClick: () => count.value++ }, 'Click me')
       )
     }
