+++ docs/examples/stack_vs_fiber/index.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html style="width: 100%; height: 100%; overflow: hidden">
+  <head>
+    <meta charset="utf-8">
+    <title>Fiber vs Stack Demo</title>
+    <link rel="stylesheet" href="base.css" />
+  </head>
+  <body>
+    <h1>Kurge Fiber vs Stack Demo</h1>
+    <div id="container">
+      <p>
+        This demo forked from <a href="https://github.com/claudiopro/react-fiber-vs-stack-demo">react-fiber-vs-stack-demo</a>
+        which is a demo of React Fiber.
+      </p>
+      <p>
+        It shows the differences between Stack and Fiber by rendering a Sierpinski triangle
+        that constantly shrinks and grows, and whose nodes have a value that increments by one each animation frame.
+      </p>
+      <ul>
+        <li><a href="fiber.html">Fiber Example</a></li>
+        <li><a href="stack.html">Stack Example</a></li>
+      </ul>
+  </body>
+</html>
