+++ docs/examples/stack_vs_fiber/kurge_1.0.5.js
