+++ docs/examples/stack_vs_fiber/kurge_1.1.0.js
