+++ docs/examples/stack_vs_fiber/stack.html
@@ -2,13 +2,13 @@
 <html style="width: 100%; height: 100%; overflow: hidden;">
   <head>
     <meta charset="utf-8">
-    <title>Stack Example</title>
+    <title>Kurge Stack Example</title>
     <link rel="stylesheet" href="base.css" />
   </head>
   <body>
-    <h1>Stack Example</h1>
+    <h1>Kurge Stack Example</h1>
     <div id="app"></div>
-    <script src="kurge.stack.js"></script>
+    <script src="kurge_1.0.5.js"></script>
     <script>
       const { render, useState, createElement: h, useEffect } = Kurge
 
@@ -65,7 +65,7 @@
         if (slowDown) {
           const e = performance.now() + 0.8
           while (performance.now() < e) {
-            // Artificially long execution time.
+            // artificially long execution time.
           }
         }
         s /= 2
@@ -95,12 +95,7 @@
         function tick() {
           seconds.value = (seconds.value % 10) + 1
         }
-        useEffect(() => {
-          const intervalID = setInterval(tick, 16)
-          return () => {
-            clearInterval(this.intervalID)
-          }
-        }, [])
+        useEffect(() => requestAnimationFrame(tick))
         
         const elapsed = Date.now() - start
         const t = (elapsed / 1000) % 10
@@ -119,10 +114,7 @@
         )
       }
 
-      render(
-        h(ExampleApplication),
-        document.getElementById('app')
-      )
+      render(h(ExampleApplication), document.getElementById('app'))
     </script>
   </body>
 </html>
