+++ types/observer/index.d.ts
@@ -1,1 +1,2 @@
-export default function observe(data: any): any;
+import Watcher from './watcher';
+export default function observe(data: any, specificWatcher?: Watcher): any;
