--- types/src/context/index.d.ts
@@ -1,1 +1,0 @@
-export default function createContext(ctx: any): any;
