--- types/src/hooks/effect.d.ts
@@ -1,2 +1,0 @@
-import { Effect } from '../shared/types';
-export default function useEffect(effect: Effect, guard?: any[]): void;
