--- types/src/hooks/refs.d.ts
@@ -1,3 +1,0 @@
-export default function useRefs(): {
-    [ref: string]: HTMLElement;
-};
