--- types/src/hooks/state.d.ts
@@ -1,1 +1,0 @@
-export default function useState(state: any): any;
