--- types/src/index.d.ts
@@ -1,16 +1,0 @@
-import render from './renderer';
-import createElement from './vdom';
-import createContext from './context';
-import useState from './hooks/state';
-import useRefs from './hooks/refs';
-import useEffect from './hooks/effect';
-declare const Kurge: {
-    version: string;
-    render: typeof render;
-    createElement: typeof createElement;
-    createContext: typeof createContext;
-    useState: typeof useState;
-    useRefs: typeof useRefs;
-    useEffect: typeof useEffect;
-};
-export default Kurge;
