--- types/src/observer/dependeny.d.ts
@@ -1,14 +1,0 @@
-import Watcher from './watcher';
-export default class Dependency {
-    static target: Watcher;
-    readonly id: number;
-    readonly list: Watcher[];
-    private readonly specificWatcher;
-    constructor(specificWatcher?: Watcher);
-    subscribe(watcher: Watcher): void;
-    unsubscribe(watcher: Watcher): void;
-    collect(): void;
-    notify(): void;
-}
-export declare function pushTarget(target: Watcher): void;
-export declare function popTarget(): void;
