--- types/src/observer/index.d.ts
@@ -1,2 +1,0 @@
-import Watcher from './watcher';
-export default function observe(data: any, specificWatcher?: Watcher): any;
