--- types/src/renderer/diff.d.ts
@@ -1,3 +1,0 @@
-import { Elem, Instance, Patches } from '../shared/types';
-export declare function diff(prevInstances: Instance[], nextChildren: Elem[]): Patches;
-export declare function patch(parentInst: Instance, patches: Patches): void;
