--- types/src/renderer/index.d.ts
@@ -1,4 +1,0 @@
-import { Elem, VDomNode } from '../shared/types';
-import { Instance } from '../shared/types';
-export declare function instantiate(element: Elem): Instance;
-export default function render(vdom: VDomNode, container: HTMLElement): void;
