--- types/src/renderer/reconciler.d.ts
@@ -1,9 +1,0 @@
-import { Elem, Instance } from '../shared/types';
-declare class Reconciler {
-    private readonly dirtyInstanceSet;
-    private isBatchUpdating;
-    enqueueUpdate(instance: Instance, element?: Elem): void;
-    private runBatchUpdate;
-}
-declare const _default: Reconciler;
-export default _default;
