--- types/src/shared/constants.d.ts
@@ -1,7 +1,0 @@
-export declare const DATA_ID: string;
-export declare const RESERVED_PROPS: {
-    [prop: string]: boolean;
-};
-export declare const SUPPORTED_LISTENERS: {
-    [listener: string]: boolean;
-};
