--- types/src/utils/dom.d.ts
@@ -1,4 +1,0 @@
-export declare function getNode(id: string): HTMLElement;
-export declare function createNode(markup: string): Text | HTMLElement;
-export declare function getClassString(className: any): string;
-export declare function getStyleString(style: any): string;
