--- types/src/vdom/index.d.ts
@@ -1,2 +1,0 @@
-import { Elem, VDomNode, Component } from '../shared/types';
-export default function createElement(type: string | Component, config?: any, ...children: (Elem | Elem[])[]): VDomNode;
