+++ types/utils/index.d.ts
@@ -1,3 +1,4 @@
+import { IdleCallback } from '../shared/types';
 export declare const is: {
     undefined: (val: any) => boolean;
     null: (val: any) => boolean;
@@ -12,7 +13,4 @@ export declare const is: {
 };
 export declare function hasOwn(object: any, property: string | number | symbol): any;
 export declare function delArrItem(arr: any[], item: any): any[] | void;
-export declare const nextTick: typeof requestAnimationFrame;
-export declare function rICB(callback: (deadline: {
-    timeRemaining: () => number;
-}) => void): any;
+export declare function nextTick(callback: IdleCallback): any;
